+++ src/components/ui/stepper/StepperText.tsx
@@ -1,4 +1,4 @@
-import { TPortfolioStepItem } from '@/model';
+import { TPortfolioStepItem } from '@/model/portfolio/portfolio';
 
 interface PStepperText {
   stepList: TPortfolioStepItem[];
