+++ src/model/portfolio/portfolio.ts
